+++ vkcs/dataplatform/datasource_product/flattens.go
@@ -35,10 +35,17 @@ func FlattenConfigs(ctx context.Context, o *products.ProductConfig) (ConfigsValu
 		return NewConfigsValueUnknown(), diags
 	}
 
+	crontabs, d := FlattenConfigsCrontabs(ctx, o.Crontabs)
+	diags.Append(d...)
+	if diags.HasError() {
+		return NewConfigsValueUnknown(), diags
+	}
+
 	configsV := ConfigsValue{
 		Connections: connections,
 		Settings:    settings,
 		UserRoles:   userRoles,
+		Crontabs:    crontabs,
 		state:       attr.ValueStateKnown,
 	}
 
@@ -193,6 +200,85 @@ func FlattenConfigsUserRoles(ctx context.Context, o []products.ProductConfigUser
 	diags.Append(d...)
 	if diags.HasError() {
 		return types.ListUnknown(ConfigsUserRolesValue{}.Type(ctx)), diags
+	}
+	return result, nil
+}
+
+func FlattenConfigsCrontabs(ctx context.Context, o []products.ProductConfigCrontabs) (basetypes.ListValue, diag.Diagnostics) {
+	var diags diag.Diagnostics
+	if o == nil {
+		return types.ListNull(ConfigsCrontabsValue{}.Type(ctx)), nil
+	}
+
+	crontabsV := make([]attr.Value, len(o))
+	for i, c := range o {
+		settings, d := FlattenConfigsCrontabsSettings(ctx, c.Settings)
+		diags.Append(d...)
+		if diags.HasError() {
+			return types.ListUnknown(ConfigsCrontabsValue{}.Type(ctx)), diags
+		}
+		crontabsV[i] = ConfigsCrontabsValue{
+			Name:     types.StringValue(c.Name),
+			Required: types.BoolValue(c.Required),
+			Start:    types.StringValue(c.Start),
+			Settings: settings,
+			state:    attr.ValueStateKnown,
+		}
+	}
+	result, d := types.ListValue(ConfigsCrontabsValue{}.Type(ctx), crontabsV)
+	diags.Append(d...)
+	if diags.HasError() {
+		return types.ListUnknown(ConfigsCrontabsValue{}.Type(ctx)), diags
+	}
+	return result, nil
+}
+
+func FlattenConfigsCrontabsSettings(ctx context.Context, o []products.ProductConfigCrontabsSettings) (basetypes.ListValue, diag.Diagnostics) {
+	var diags diag.Diagnostics
+	if o == nil {
+		return types.ListNull(ConfigsCrontabsSettingsValue{}.Type(ctx)), nil
+	}
+
+	settingsV := make([]attr.Value, len(o))
+	for i, s := range o {
+		stringVariation, d := FlattenConfigsCrontabsSettingsStringVariation(ctx, s.StringVariation)
+		diags.Append(d...)
+		if diags.HasError() {
+			return types.ListUnknown(ConfigsCrontabsSettingsValue{}.Type(ctx)), diags
+		}
+
+		settingsV[i] = ConfigsCrontabsSettingsValue{
+			Alias:           types.StringValue(s.Alias),
+			DefaultValue:    types.StringValue(s.DefaultValue),
+			IsRequire:       types.BoolValue(s.IsRequired),
+			IsSensitive:     types.BoolValue(s.IsSensitive),
+			Regexp:          types.StringValue(s.RegExp),
+			StringVariation: stringVariation,
+			state:           attr.ValueStateKnown,
+		}
+	}
+	result, d := types.ListValue(ConfigsCrontabsSettingsValue{}.Type(ctx), settingsV)
+	diags.Append(d...)
+	if diags.HasError() {
+		return types.ListUnknown(ConfigsCrontabsSettingsValue{}.Type(ctx)), diags
+	}
+	return result, nil
+}
+
+func FlattenConfigsCrontabsSettingsStringVariation(ctx context.Context, o []string) (basetypes.ListValue, diag.Diagnostics) {
+	var diags diag.Diagnostics
+	if o == nil {
+		return types.ListNull(types.StringType), nil
+	}
+
+	settingsV := make([]attr.Value, len(o))
+	for i, s := range o {
+		settingsV[i] = types.StringValue(s)
+	}
+	result, d := types.ListValue(types.StringType, settingsV)
+	diags.Append(d...)
+	if diags.HasError() {
+		return types.ListUnknown(types.StringType), diags
 	}
 	return result, nil
 }