+++ CHANGELOG.md
@@ -8,6 +8,7 @@ description: |-
 # VKCS Provider's changelog
 #### v0.13.2 (unreleased)
 - Fix bug with missing error details when instance creation fails
+- Add crontabs to vkcs_dataplatform_product
 
 #### v0.13.1
 - Support adding and removing connections for vkcs_dataplatform_cluster